Added input and output methods for running the program.
There are now two input gates available. `Clock` which will send out a HIGH signal every time. And `AutomaticInput` which allows me to put a vector of signals in, and they will be called in order.

There is also an output gate called `SimpleOutput`. The results of output gates will be printed every time a clock tick completes. This will allow for the circuit to stop partway (because there were no more changes to make) and the program will still print the output.