Refactored the main function out into build_simple_circuits and run_circuit. This should allow me to easily drop in and save multiple functions that let me build different circuits. It also allows me to call run_circuit to test different gates.
Fixed bug where signal would not always propagate to the output when the input did not change. This was set up so that on the first clock tick, it will always go through the entire circuit so that the output is correct.

Added the basic logic gates in And, Nor and Nand.

Added in tests for simple gates as well as a few pitfalls when running the circuit.