Fixed a function in the infrastructure of the simulation. What was happening was the NONE calculation was not working correctly if the gate ran multiple times on the current clock tick.
aTwo changes were made to fix this. First the check for if it was the first clock tick was removed from update_input_signal. Second, if multiple values feed into the same input, the error will be handled by run_circuit. It will put the gate in a separate vector and will wait until all other gates have been properly run to make sure that the proper input was not pending.

Added an optimization to run_circuit was added. If a group of gates should be all run at the same time, that type can now be passed to run_circuit.

The RAMUnit is now completed. Next is the arithmetic logic unit.