Trying to figure out the deadlock problems here. Still can't get the deadlock problem. The Mutex are not actually locked inside each other, so the conventional way of deadlocking should not be possible.
Changed the program so that it never holds a lock, it only locks and immediately drops it. Made sure to manually catch panic results in case any of them are missed.

Neither of these things illuminated the situation. I am starting to think that there is some king of problem with how Rust works with the OS or hardware. I might want to try running this program on Windows or a different version of Linux and see if I have more success.